% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice_2562.R
\docType{data}
\name{rice_2562}
\alias{rice_2562}
\title{Thai major rice production of 2562//63 (2019/20) crop year by province}
\format{
An object of class \code{"sf"}
}
\source{
\href{http://www.oae.go.th}{}
}
\usage{
data(rice_2562)
}
\description{
Data from Office of Agricultural Economics (OAE)
}
\examples{
data(rice_2562)
head(rice_2562)
}
\keyword{datasets}
